Move additional services on product to separate property
According to the mybring API additional services on a product should be a child element with name services.

This commit movies the RecipientNotification into the Services class. Also add DeliveryOption property.

The properties on Services are mutable, as they correctly initialize to null, but can be sett on an individual basis based on which additional services you want to enable.